requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
boto3==1.37.13
botocore==1.37.13
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==25.1
django-hosts==6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycryptodome==3.22.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.11.4
sentry-sdk==2.22.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid7==0.1.0
whitenoise==6.9.0
freezegun==1.5.1
better-profanity==0.7.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
celery==5.5.0
django-redis==5.4.0
psycopg2-binary==2.9.10